MAINT: move neuron -> neuron_network
