+++ hnn_core/viz.py
@@ -131,8 +131,8 @@ def plot_cells(net, ax=None, show=True):
 
     Parameters
     ----------
-    net : instance of NeuronNetwork
-        The NeuronNetwork object.
+    net : instance of NetworkBuilder
+        The NetworkBuilder object.
     ax : instance of matplotlib Axes3D | None
         An axis object from matplotlib. If None,
         a new figure is created.
